refactor: legacy 코드 주석 처리 및 제거
