+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,15 +1,13 @@
 package com.techcourse;
 
-import com.techcourse.controller.*;
+import java.util.HashMap;
+import java.util.Map;
+
 import jakarta.servlet.http.HttpServletRequest;
 import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.asis.ForwardController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class ManualHandlerMapping implements HandlerMapping {
 
@@ -19,15 +17,17 @@ public class ManualHandlerMapping implements HandlerMapping {
 
     @Override
     public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
-        controllers.put("/login", new LoginController());
-        controllers.put("/login/view", new LoginViewController());
-        controllers.put("/logout", new LogoutController());
-        controllers.put("/register/view", new RegisterViewController());
-        controllers.put("/register", new RegisterController());
+        // as-is에 대한 제거를 위한 주석
+        //        controllers.put("/", new ForwardController("/index.jsp"));
+        //        controllers.put("/login", new LoginController());
+        //        controllers.put("/login/view", new LoginViewController());
+        //        controllers.put("/logout", new LogoutController());
+        //        controllers.put("/register/view", new RegisterViewController());
+        //        controllers.put("/register", new RegisterController());
 
         log.info("Initialized Handler Mapping!");
-        controllers.keySet().forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
+        controllers.keySet()
+                   .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
     }
 
     @Override
