+++ app/src/main/java/com/techcourse/controller/asis/LoginController.java
@@ -1,5 +1,6 @@
-package com.techcourse.controller;
+package com.techcourse.controller.asis;
 
+import com.techcourse.controller.UserSession;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
