+++ app/src/main/java/com/techcourse/controller/asis/LoginViewController.java
@@ -1,4 +1,6 @@
-package com.techcourse.controller;
+package com.techcourse.controller.asis;
+
+import com.techcourse.controller.UserSession;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -13,10 +15,10 @@ public class LoginViewController implements Controller {
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
         return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
+                          .map(user -> {
                     log.info("logged in {}", user.getAccount());
                     return "redirect:/index.jsp";
                 })
-                .orElse("/login.jsp");
+                          .orElse("/login.jsp");
     }
 }
