+++ app/src/main/java/com/techcourse/controller/asis/LogoutController.java
@@ -1,4 +1,6 @@
-package com.techcourse.controller;
+package com.techcourse.controller.asis;
+
+import com.techcourse.controller.UserSession;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
